update module name on stringifier
